Use ffmpeg to automatically get podcast duration (#1)

+++ rss_hosting/deploy_podcasts.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 echo $1
+mkdir -p public/
 timestamp=$(date +%y%m%d_%H%M%S)
-mv public/rss.xml public/rss_$timestamp.xml
-rss_sample=$(cat $1/feed.sample.xml)
+[ ! -f public/rss.xml ] || mv public/rss.xml public/rss_$timestamp.xml
+rss_sample=$(./fill_podcast_duration.sh $1)
 echo "
 <rss version=\"2.0\" xmlns:itunes=\"http://www.itunes.com/dtds/podcast-1.0.dtd\">
     <script />
@@ -17,6 +18,8 @@ echo "
 </rss>
 " > public/rss.xml
 
-cp $1/*.mp3 public
-firebase deploy
-echo "Now go on https://www.playrun.app/podcast and add the private podcast with the url https://podcasts-noan.web.app/rss.xml"
+
+
+# cp $1/*.mp3 public
+# firebase deploy
+# echo "Now go on https://www.playrun.app/podcast and add the private podcast with the url https://podcasts-noan.web.app/rss.xml"
